refactor(reviewer-assign): split comms_delegate into a directory module to satisfy file_size_invariant — no behavior change (t-…-17 slice 4)
comms_delegate.rs grew 679→1374 LOC (reviewer-assignment gate + ~600 LOC of
inline tests), breaching the 750-LOC file_size_invariant cap. Convert it to a
directory module: move the two gate fns (validate_review_assignment_marker,
dispatch_review_assignment_via_store) verbatim into review_assignment.rs and
both inline test modules into the gate-exempt tests.rs. mod.rs is now 562 LOC.
Pure relocation — production fn bodies are byte-identical; only visibility
(pub(super)) and import paths changed. comms.rs #[path] retargeted to mod.rs.


Agend-Agent: archfix-opus-2
Agend-Task: t-20260713002749179875-15764-17
Agend-Branch: fix/reviewer-assignment-outbox
Agend-Issued-At: 2026-07-13T00:35:43.419484+00:00
